Support aiobotocore version 1.0
In particular we need to use aiobotocore clients in an async context
manager.  So now rather than store these we recreate them each time.
This adds a modest cost increase of about 5ms on my machine, but also
simplifies cleanup a bit.

I believe that this change is also backwards compatibile

Fixes #80
Fixes #78