Avoid deprecated method to get subscription ID
This removes the use of a deprecated method from azure-cli-core
to get the default subscription ID. It also removes the dependency
on azure-cli-core.

For the common case, where the user has the Azure CLI installed and
configured, we can get the subscription ID by parsing its output.

This PR also enables users to configure the subscription ID either
in code or through the Dask config system.

A few related documentation updates for good measure.